select-list-view: Improve key detection to work in Chrome/Safari and Firefox.
Switch from using KeyIdentifier to KeyCode.
The keyboard event API seems to be a mess. KeyIdentifier works in
Chrome/Safari, but isn't present in Firefox. MDN docs say KeyCode is
deprecated and Key should be used instead, but Chrome/Safari do not
have Key set. KeyCode seems to be the only solution that worked in all
3 browsers (latest stable versions). I'll go with that for now, but
this seems like a great place to create a library that abstracts out
the browser idiosyncrasies.
This should help with shurcooL/gtdo#21.